Moving WrapAsync code into AsyncValueFn as older versions of python don't support [] to define ParamSpec
