+++ shiny/render/transformer/_transformer.py
@@ -32,7 +32,7 @@ from typing import (
     overload,
 )
 
-from ..renderer import JSONifiable, RendererBase, WrapAsync
+from ..renderer import AsyncValueFn, JSONifiable, RendererBase
 from ..renderer._renderer import DefaultUIFn, DefaultUIFnResultOrNone
 
 if TYPE_CHECKING:
@@ -260,7 +260,7 @@ class OutputRenderer(RendererBase, Generic[OT]):
         # Checking if a function is async has a 180+ns overhead (barret's machine)
         # -> It is faster to always call an async function than to always check if it is async
         # Always being async simplifies the execution
-        self._value_fn = WrapAsync(value_fn)
+        self._value_fn = AsyncValueFn(value_fn)
         self._value_fn_is_async = self._value_fn.is_async  # legacy key
         self.__name__ = value_fn.__name__
